Moved unit tests into their correct test file.
